+++ lib/Processor/MPDMSE/SourceBufferTaskCollection.ts
@@ -8,12 +8,12 @@ import { parseRepresentationElement } from "./Representation";
 
 /** 更新 SourceBufferTasks */
 const updateSourceBufferTasks = (mpdElement: Element, taskCollection: SourceBufferTaskCollection): void => {
-    taskCollection.duration = parsePTdurationToSeconds(mpdElement.getAttribute("mediaPresentationDuration"),Infinity);
+    taskCollection.duration = parsePTdurationToSeconds(mpdElement.getAttribute("mediaPresentationDuration"), Infinity);
     taskCollection.clear();
     for (const adaptationSetElement of mpdElement.getElementsByTagName("AdaptationSet")) {
         const contentType = adaptationSetElement.getAttribute("contentType");
         if (contentType) {
-            const sourceBufferTask = taskCollection.getORset(contentType)
+            const sourceBufferTask = taskCollection.set(contentType)
             for (const repElement of adaptationSetElement.getElementsByTagName("Representation")) {
                 sourceBufferTask.set(
                     parseRepresentationElement(repElement, taskCollection.duration, adaptationSetElement.getElementsByTagName("SegmentList")?.[0])
@@ -26,7 +26,7 @@ const updateSourceBufferTasks = (mpdElement: Element, taskCollection: SourceBuff
 
 /** 源缓存任务映射集合 */
 export class SourceBufferTaskCollection {
-   // #options: PlayerOptions
+    // #options: PlayerOptions
     #map = new Map<RepType, SourceBufferTask>();
     #mse = new MediaSource();
     #mpdConverter: MPDDefaultConverter;
@@ -36,30 +36,29 @@ export class SourceBufferTaskCollection {
     constructor(mpdConverter: MPDDefaultConverter, el: HTMLMediaElement, options: PlayerOptions, eventEmitter: PlayerEventEmitter) {
         this.#mpdConverter = mpdConverter
         this.#el = el;
-       // this.#options = options
+        // this.#options = options
         this.#eventEmitter = eventEmitter
-        this.#fetchScheduleFactoryMethod = createFetchScheduleFactoryMethod(mpdConverter,options)
+        this.#fetchScheduleFactoryMethod = createFetchScheduleFactoryMethod(mpdConverter, options)
 
     }
     get duration() { return this.#mse.duration }
     set duration(val: number) { this.#mse.duration = val || this.#mse.duration }
     sourceBufferUpdate(currentTime: number) {
-        if (this.#mse.readyState === "open") {
-        //    console.log("SourceBufferTaskCollection -sourceBufferUpdate",currentTime);
-            Promise.all(this.#map.values().map(s => s.sourceBufferUpdate(currentTime))).then(sourceBufferTasks => {
-                if (sourceBufferTasks.every(sourceBufferTask => sourceBufferTask.isLastFile(currentTime))) {
-                    if (Number.isInteger(this.#mse.duration)) {
-                        if (this.#mse.readyState === "open") { this.#mse.endOfStream() }
-                    } else {
-                        this.#mpdConverter.asyncResponse().then(async response => {
-                            if (response && response.ok) {
-                                const mpdElement = new DOMParser().parseFromString(await response.text(), "text/xml").documentElement;
-                                updateSourceBufferTasks(mpdElement, this)
-                            }
-                        })
-                    }
+        if (this.#mse.readyState === "open") { 
+            if (Array.from(this.#map.values()).every(sourceBufferTask =>
+                sourceBufferTask.isLastFile(currentTime) ||!(sourceBufferTask.sourceBufferUpdate(currentTime)) )) {
+                if (Number.isInteger(this.#mse.duration)) {
+                    if (this.#mse.readyState === "open") { this.#mse.endOfStream() }
+                } else {
+                    this.#mpdConverter.asyncResponse().then(async response => {
+                        if (response && response.ok) {
+                            const mpdElement = new DOMParser().parseFromString(await response.text(), "text/xml").documentElement;
+                            updateSourceBufferTasks(mpdElement, this)
+                        }
+                    })
                 }
-            })
+            }
+
         }
     }
 
@@ -68,24 +67,30 @@ export class SourceBufferTaskCollection {
         this.#map.values().forEach(s => s.clear())
         return this
     }
-    /** 获取或者设置 SourceBufferTask*/
-    getORset(repType: RepType) {
+    /** 设置 SourceBufferTask */
+    set(repType: RepType) {
         return this.#map.has(repType) ?
             this.#map.get(repType)!
             : this.#map.set(repType, new SourceBufferTask(this.#mse, this.#fetchScheduleFactoryMethod)).get(repType)!
     }
-    changeType(repType: RepType, rep: Representation,  currentTime:number,options: SwitchRepOptions) {
-        this.#map.get(repType)?.switch(rep,currentTime, options)
+    /** 获取 SourceBufferTask*/
+    get(repType: RepType) {
+        return this.#map.has(repType) ?
+            this.#map.get(repType)!
+            : undefined
+    }
+    changeType(repType: RepType, rep: Representation, currentTime: number, options: SwitchRepOptions) {
+        this.#map.get(repType)?.switch(rep, currentTime, options)
     }
     asyncSourceopen(res?: Response) {
-        const { promise, resolve } = Promise.withResolvers<typeof this>()      
-        this.#mse.addEventListener("sourceopen",async () => {    
-            const response = res instanceof Response ? res : (await  this.#mpdConverter.asyncResponse())
-                if (response && response.ok) {
-                    const mpdElement = new DOMParser().parseFromString(await response.text(), "text/xml").documentElement;
-                    updateSourceBufferTasks(mpdElement, this)
-                }
-                resolve(this)
+        const { promise, resolve } = Promise.withResolvers<typeof this>()
+        this.#mse.addEventListener("sourceopen", async () => {
+            const response = res instanceof Response ? res : (await this.#mpdConverter.asyncResponse())
+            if (response && response.ok) {
+                const mpdElement = new DOMParser().parseFromString(await response.text(), "text/xml").documentElement;
+                updateSourceBufferTasks(mpdElement, this)
+            }
+            resolve(this)
         }, { once: true })
         this.#el.src = URL.createObjectURL(this.#mse);
         return promise