+++ lib/Processor/MPDMSE/index.ts
@@ -1,18 +1,18 @@
 import { PlayerOptions, PlayerEventEmitter, SwitchRepOptions } from "@lib"
-import type { Representation, Processor, ProcessorType, RepType } from "@lib"
+import type { Representation, Processor, ProcessorFactory, RepType } from "@lib"
 import { MPDDefaultConverter, MPDConverter } from "./Converter"
 import { SourceBufferTaskCollection } from "./SourceBufferTaskCollection"
-export { type MPDConverter }
+
 /**
  * MPD MSE媒体源处理器
  */
-class MPDMSE implements Processor {
+class MPDMSEProcessor implements Processor {
     #SourceBufferTaskCollection: SourceBufferTaskCollection
     get(repType: RepType) {
-        return this.#SourceBufferTaskCollection.getORset(repType).toArray()
+        return this.#SourceBufferTaskCollection.get(repType)?.toArray() ?? []
     }
-    switch(repType: RepType, rep: Representation, currentTime:number, options: SwitchRepOptions) { 
-        this.#SourceBufferTaskCollection.changeType(repType, rep,currentTime, options)
+    switch(repType: RepType, rep: Representation, currentTime: number, options: SwitchRepOptions) {
+        this.#SourceBufferTaskCollection.changeType(repType, rep, currentTime, options)
     }
     sourceBufferUpdate(currentTime: number) { this.#SourceBufferTaskCollection.sourceBufferUpdate(currentTime) }
     constructor(sourceBufferTaskCollection: SourceBufferTaskCollection) {
@@ -21,21 +21,21 @@ class MPDMSE implements Processor {
 
 }
 
-const processor: ProcessorType = {
+const MPDMSEFactory: ProcessorFactory = {
     name: "MPDMSE",
-    asyncFunctionProcessorInstance: async (result: unknown, el: HTMLMediaElement, options: PlayerOptions, eventEmitter: PlayerEventEmitter) => {
+    asyncCreateProcessorInstance: async (result: unknown, el: HTMLMediaElement, options: PlayerOptions, eventEmitter: PlayerEventEmitter) => {
         const mpdConverter = MPDDefaultConverter.parse(result instanceof Response ? result.url : result)
         if (mpdConverter instanceof MPDDefaultConverter) {
-            return mpdConverter.asyncResponse().then(response => { 
+            return mpdConverter.asyncResponse().then(response => {
                 return response.ok ? (new SourceBufferTaskCollection(mpdConverter, el, options, eventEmitter)).asyncSourceopen(response).then(s => {
-                   // console.log("mpdmse",s,response);
-                   // console.log(response);
-                    return new MPDMSE(s)
-                }).catch(()=>undefined) : undefined
+                    // console.log("mpdmse",s,response);
+                    // console.log(response);
+                    return new MPDMSEProcessor(s)
+                }).catch(() => undefined) : undefined
             })
         }
         return undefined
     }
 }
-export default processor
 
+export { MPDMSEFactory, type MPDConverter }